+++ docs/changes.rst
@@ -30,6 +30,8 @@ Not yet released.
 
 **Compatibility**
 
+* :ref:`rollbar-errors` integration no longer includes client-side error collection.
+
 **Upgrading**
 
 Please follow :ref:`generic-upgrade-instructions` in order to perform update.